Feature login

+++ src/main/java/net/den3/den3Account/Store/Auth/AuthorizationStore.java
@@ -21,19 +21,17 @@ public class AuthorizationStore implements IAuthorizationStore{
     /**
      * ストアにサービス認可済みアカウントのテーブルを作る
      *
-     * @param service
      * @return true->成功 false->失敗
      */
     @Override
-    public boolean createTableServiceAccount(IService service) {
-        if(hasTableServiceAccount(service)){
+    public boolean createTableServiceAccount() {
+        if(hasTableServiceAccount()){
             //すでにテーブルが存在している
             return false;
         }
         return store.controlSQL(con->{
             try{
-                PreparedStatement ps = con.prepareStatement("create table ? (uuid VARCHAR(255));");
-                ps.setString(1,"service_"+service.getServiceID());
+                PreparedStatement ps = con.prepareStatement("create table authorization_store ( uuid VARCHAR(255) not null primary key, service_id VARCHAR(255) );");
                 return Optional.of(Collections.singletonList(ps));
             }catch (SQLException ex){
                 return Optional.empty();
@@ -44,17 +42,15 @@ public class AuthorizationStore implements IAuthorizationStore{
     /**
      * ストアがサービス認可済みアカウントのテーブルを持っているか
      *
-     * @param service
      * @return true->持っている false->持っていない
      */
     @Override
-    public boolean hasTableServiceAccount(IService service) {
+    public boolean hasTableServiceAccount() {
         AtomicBoolean b = new AtomicBoolean(true);
         List<String> columns = Collections.singletonList("uuid");
         store.getLineBySQL(columns,(con)->{
             try{
-                PreparedStatement ps = con.prepareStatement("SELECT * FROM ?");
-                ps.setString(1,"service_"+service.getServiceID());
+                PreparedStatement ps = con.prepareStatement("SELECT * FROM authorization_store");
                 return Optional.of(ps);
             }catch (SQLException ex){
                 b.set(false);
@@ -64,24 +60,24 @@ public class AuthorizationStore implements IAuthorizationStore{
         return b.get();
     }
 
+
     /**
-     * アカウントがアプリに個人情報の使用を認可しているかどうか
-     *
-     * @param account 調べる対象のアカウント
-     * @param service 調べるアカウント
-     * @return true->認可済み false->未認可
+     * アカウントをアプリの個人情報使用認可ストアに追加する
+     * @param account 調べる対象のアカウントのUUID
+     * @param service 調べるサービスのUUID
+     * @return
      */
     @Override
-    public boolean isUserAuthorization(IAccount account, IService service) {
-        if(!hasTableServiceAccount(service)){
+    public boolean isUserAuthorization(String account, String service) {
+        if(!hasTableServiceAccount()){
             return false;
         }
-        List<String> columns = Collections.singletonList("uuid");
+        List<String> columns = Arrays.asList("uuid","service_name");
         Optional<List<Map<String, String>>> results = store.getLineBySQL(columns,con -> {
             try {
-                PreparedStatement ps = con.prepareStatement("SElECT * FROM ? WHERE uuid = ?");
-                ps.setString(1, "service_" + service.getServiceID());
-                ps.setString(2, account.getUUID());
+                PreparedStatement ps = con.prepareStatement("SElECT * FROM authorization_store WHERE uuid = ? and service_id = ?");
+                ps.setString(1, account);
+                ps.setString(2, service);
                 return Optional.of(ps);
             } catch (SQLException ex) {
                 ex.printStackTrace();
@@ -100,12 +96,12 @@ public class AuthorizationStore implements IAuthorizationStore{
      * @return true->成功 false->失敗
      */
     @Override
-    public boolean addAuthorizationUser(IAccount account, IService service) {
+    public boolean addAuthorizationUser(String account, String service) {
         return store.controlSQL(con->{
             try{
-                PreparedStatement ps = con.prepareStatement("INSERT INTO ? VALUES (?) ;");
-                ps.setString(1,"service_" + service.getServiceID());
-                ps.setString(2,account.getUUID());
+                PreparedStatement ps = con.prepareStatement("INSERT INTO authorization_store VALUES (?,?) ;");
+                ps.setString(1,account);
+                ps.setString(2,service);
                 return Optional.of(Collections.singletonList(ps));
             }catch (SQLException ex){
                 ex.printStackTrace();
@@ -117,17 +113,17 @@ public class AuthorizationStore implements IAuthorizationStore{
     /**
      * アカウントをアプリの個人情報使用認可ストアから削除する
      *
-     * @param account 追加するアカウント
-     * @param service 削除先のサービス
+     * @param account 削除するアカウントのUUID
+     * @param service 削除先のサービスのUUID
      * @return true->成功 false->失敗
      */
     @Override
-    public boolean deleteAuthorizationUser(IAccount account, IService service) {
+    public boolean deleteAuthorizationUser(String account, String service) {
         return store.controlSQL(con->{
             try{
-                PreparedStatement ps = con.prepareStatement("DELETE FROM ? WHERE uuid = ?");
-                ps.setString(1,"service_" + service.getServiceID());
-                ps.setString(2,account.getUUID());
+                PreparedStatement ps = con.prepareStatement("DELETE FROM authorization_store WHERE uuid = ? and service_id = ?");
+                ps.setString(1,account);
+                ps.setString(2,service);
                 return Optional.of(Collections.singletonList(ps));
             }catch (SQLException ex){
                 ex.printStackTrace();
@@ -136,5 +132,42 @@ public class AuthorizationStore implements IAuthorizationStore{
         });
     }
 
+    /**
+     * アカウントをアプリの個人情報使用認可ストアから削除する
+     *
+     * @param account 削除するアカウントのUUID
+     * @return true->成功 false->失敗
+     */
+    @Override
+    public boolean deleteAuthorizationUser(String account) {
+        return store.controlSQL(con->{
+            try{
+                PreparedStatement ps = con.prepareStatement("DELETE FROM authorization_store WHERE uuid = ?");
+                ps.setString(1,account);
+                return Optional.of(Collections.singletonList(ps));
+            }catch (SQLException ex){
+                ex.printStackTrace();
+                return Optional.empty();
+            }
+        });
+    }
 
+    /**
+     * サービスに紐づけられたアカウントをすべて削除する
+     * @param service 対象のサービス
+     * @return true->成功 false->失敗
+     */
+    @Override
+    public boolean deleteAuthorizationUser(IService service) {
+        return store.controlSQL(con->{
+            try{
+                PreparedStatement ps = con.prepareStatement("DELETE FROM authorization_store WHERE service_id = ?");
+                ps.setString(1,service.getServiceID());
+                return Optional.of(Collections.singletonList(ps));
+            }catch (SQLException ex){
+                ex.printStackTrace();
+                return Optional.empty();
+            }
+        });
+    }
 }
